RavenDB-20467 Syntax improvements
